pyproject.toml
[tool.poetry]
name = "simple_homepage"
version = "0.0.4"
description = "Create a simple homepage"
authors = ["Florian Maas <ffpgmaas@gmail.com>"]
repository = "https://github.com/fpgmaas/simple-homepage"
documentation = "https://fpgmaas.github.io/simple-homepage/"
readme = "README.md"
packages = [
  {include = "simple_homepage"}
]
include = ['simple_homepage/files/*']

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
oyaml = "^1.0"


[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest = "^7.1.1"
mkdocs = "^1.3.0"
mkdocs-material = "^8.2.9"
mkdocstrings = "^0.18.1"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
target-version = ['py39']
fast = true
exclude = '''
(
  /(                        # exclude a few common directories in the
    \.git                   # root of the project
    | \.pytest_cache
    | python-venv
    | \.venv
    | build
    | dist
    | \.tox
  ))
'''

[tool.isort]
profile = "black"

[tool.mypy]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
exclude = [
    '\.venv',
    'tests'
]

[tool.poetry.scripts]
homepage = "simple_homepage.cli:cli"
